# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "thewiki API",
    "description": "REST endpoints for thewiki — see https://github.com/i-doll/thewiki",
    "contact": {
      "name": "thewiki contributors"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/admin/blocklist/ip": {
      "get": {
        "tags": [
          "admin-blocklist"
        ],
        "summary": "`GET /api/v1/admin/blocklist/ip` — list every IP blocklist row.",
        "operationId": "list_ip",
        "responses": {
          "200": {
            "description": "IP blocklist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpBlocklistListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks MANAGE_BLOCKLIST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin-blocklist"
        ],
        "summary": "`POST /api/v1/admin/blocklist/ip` — add a new IP blocklist row.",
        "operationId": "create_ip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIpEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Row created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpBlocklistView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid CIDR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks MANAGE_BLOCKLIST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "CIDR already in list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/admin/blocklist/ip/{id}": {
      "delete": {
        "tags": [
          "admin-blocklist"
        ],
        "summary": "`DELETE /api/v1/admin/blocklist/ip/{id}` — remove a blocklist row.",
        "operationId": "delete_ip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Row identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Row removed"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks MANAGE_BLOCKLIST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Row not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/admin/blocklist/url": {
      "get": {
        "tags": [
          "admin-blocklist"
        ],
        "summary": "`GET /api/v1/admin/blocklist/url` — list every URL blocklist row.",
        "operationId": "list_url",
        "responses": {
          "200": {
            "description": "URL blocklist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlBlocklistListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks MANAGE_BLOCKLIST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin-blocklist"
        ],
        "summary": "`POST /api/v1/admin/blocklist/url` — add a new URL blocklist row.",
        "operationId": "create_url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUrlEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Row created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlBlocklistView"
                }
              }
            }
          },
          "400": {
            "description": "Invalid regex",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks MANAGE_BLOCKLIST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Pattern already in list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
    "/api/v1/admin/blocklist/url/{id}": {
      "delete": {
        "tags": [
          "admin-blocklist"
        ],
        "summary": "`DELETE /api/v1/admin/blocklist/url/{id}` — remove a blocklist row.",
        "operationId": "delete_url",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Row identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Row removed"
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Caller lacks MANAGE_BLOCKLIST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Row not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIpEntryRequest": {
        "type": "object",
        "description": "Body for `POST /api/v1/admin/blocklist/ip`.",
        "required": [
          "cidr"
        ],
        "properties": {
          "cidr": {
            "type": "string",
            "description": "CIDR in canonical form. The handler parses with `ipnet::IpNet::from_str`\nbefore persisting — invalid input returns 400 without touching the DB."
          },
          "reason": {
            "type": "string",
            "description": "Optional reason. Omitting it stores an empty string."
          }
        }
      },
...
      "CreateUrlEntryRequest": {
        "type": "object",
        "description": "Body for `POST /api/v1/admin/blocklist/url`.",
        "required": [
          "pattern"
        ],
        "properties": {
          "pattern": {
            "type": "string",
            "description": "Rust `regex` pattern. Compiled with `regex::Regex::new` for validation\nbefore persisting."
          },
          "reason": {
            "type": "string",
            "description": "Optional reason."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The wire form of an API error.\n\nKept intentionally small: a stable machine-readable `code` plus a free-form\nhuman message. New variants can be added by extending [`ApiError`] without\nchanging this shape.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code (`not_found`, `conflict`, …)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of what went wrong."
          }
        }
      },
      "IpBlocklistListResponse": {
        "type": "object",
        "description": "Wrapper response for `GET` endpoints.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpBlocklistView"
            },
            "description": "Rows, newest first."
          }
        }
      },
      "IpBlocklistView": {
        "type": "object",
        "description": "IP blocklist row as returned by the API.",
        "required": [
          "id",
          "cidr",
          "reason",
          "created_by",
          "created_at"
        ],
        "properties": {
          "cidr": {
            "type": "string",
            "description": "CIDR in canonical human form (`203.0.113.0/24`, `2001:db8::/32`)."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 timestamp."
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "description": "User who created the entry."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable row identifier (UUIDv7)."
          },
          "reason": {
            "type": "string",
            "description": "Free-form reason (may be empty)."
          }
        }
      },
...
      "RateLimitErrorBody": {
        "type": "object",
        "description": "Wire form for a 429 response body.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable error code. Always `\"rate_limited\"`."
          }
        }
      },
...
      "UrlBlocklistListResponse": {
        "type": "object",
        "description": "Wrapper response for `GET` endpoints.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UrlBlocklistView"
            },
            "description": "Rows, newest first."
          }
        }
      },
      "UrlBlocklistView": {
        "type": "object",
        "description": "URL blocklist row as returned by the API.",
        "required": [
          "id",
          "pattern",
          "reason",
          "created_by",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "RFC3339 timestamp."
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "description": "User who created the entry."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable row identifier (UUIDv7)."
          },
          "pattern": {
            "type": "string",
            "description": "Rust `regex` pattern."
          },
          "reason": {
            "type": "string",
            "description": "Free-form reason (may be empty)."
          }
        }
      },
...
    },
    "securitySchemes": {
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "Double-submit CSRF token matching the thewiki_csrf cookie."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "thewiki_session",
        "description": "Opaque session cookie issued by POST /api/v1/auth/login."
      }
    }
  },
...
}